migrate_tensor_array_to_tensor
